--- cmd/archive/release/runner.go
@@ -1,43 +1,0 @@
-package release
-
-import (
-	"context"
-	"io"
-
-	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/micrologger"
-	"github.com/spf13/cobra"
-
-	"github.com/giantswarm/devctl/pkg/release"
-)
-
-type runner struct {
-	flag   *flag
-	logger micrologger.Logger
-	stdout io.Writer
-	stderr io.Writer
-}
-
-func (r *runner) Run(cmd *cobra.Command, args []string) error {
-	ctx := context.Background()
-
-	err := r.flag.Validate()
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	err = r.run(ctx, cmd, args)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	return nil
-}
-
-func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) error {
-	err := release.ArchiveRelease(r.flag.Name, r.flag.Releases, r.flag.Provider)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-	return nil
-}
