+++ cmd/command.go
@@ -8,10 +8,9 @@ import (
 	"github.com/giantswarm/micrologger"
 	"github.com/spf13/cobra"
 
-	"github.com/giantswarm/devctl/cmd/archive"
 	"github.com/giantswarm/devctl/cmd/completion"
-	"github.com/giantswarm/devctl/cmd/create"
 	"github.com/giantswarm/devctl/cmd/gen"
+	"github.com/giantswarm/devctl/cmd/release"
 	"github.com/giantswarm/devctl/cmd/replace"
 	"github.com/giantswarm/devctl/cmd/repo"
 	"github.com/giantswarm/devctl/cmd/version"
@@ -52,20 +51,6 @@ func New(config Config) (*cobra.Command, error) {
 
 	var err error
 
-	var archiveCmd *cobra.Command
-	{
-		c := archive.Config{
-			Logger: config.Logger,
-			Stderr: config.Stderr,
-			Stdout: config.Stdout,
-		}
-
-		archiveCmd, err = archive.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
 	var completionCmd *cobra.Command
 	{
 		c := completion.Config{
@@ -80,20 +65,6 @@ func New(config Config) (*cobra.Command, error) {
 		}
 	}
 
-	var createCmd *cobra.Command
-	{
-		c := create.Config{
-			Logger: config.Logger,
-			Stderr: config.Stderr,
-			Stdout: config.Stdout,
-		}
-
-		createCmd, err = create.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
 	var genCmd *cobra.Command
 	{
 		c := gen.Config{
@@ -103,6 +74,20 @@ func New(config Config) (*cobra.Command, error) {
 		}
 
 		genCmd, err = gen.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
+	var releaseCmd *cobra.Command
+	{
+		c := release.Config{
+			Logger: config.Logger,
+			Stderr: config.Stderr,
+			Stdout: config.Stdout,
+		}
+
+		releaseCmd, err = release.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -172,13 +157,12 @@ func New(config Config) (*cobra.Command, error) {
 
 	f.Init(c)
 
-	c.AddCommand(archiveCmd)
 	c.AddCommand(completionCmd)
-	c.AddCommand(createCmd)
 	c.AddCommand(genCmd)
-	c.AddCommand(versionCmd)
+	c.AddCommand(releaseCmd)
 	c.AddCommand(repoCmd)
 	c.AddCommand(replaceCmd)
+	c.AddCommand(versionCmd)
 
 	return c, nil
 }
