+++ cmd/release/command.go
@@ -7,11 +7,14 @@ import (
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	"github.com/spf13/cobra"
+
+	"github.com/giantswarm/devctl/cmd/release/archive"
+	"github.com/giantswarm/devctl/cmd/release/create"
 )
 
 const (
 	name        = "release"
-	description = "Archives and de-registers an existing Giant Swarm platform release."
+	description = "Commands for working with releases on the local filesystem."
 )
 
 type Config struct {
@@ -31,6 +34,36 @@ func New(config Config) (*cobra.Command, error) {
 		config.Stdout = os.Stdout
 	}
 
+	var err error
+
+	var archiveCmd *cobra.Command
+	{
+		c := archive.Config{
+			Logger: config.Logger,
+			Stderr: config.Stderr,
+			Stdout: config.Stdout,
+		}
+
+		archiveCmd, err = archive.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
+	var createCmd *cobra.Command
+	{
+		c := create.Config{
+			Logger: config.Logger,
+			Stderr: config.Stderr,
+			Stdout: config.Stdout,
+		}
+
+		createCmd, err = create.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
 	f := &flag{}
 
 	r := &runner{
@@ -48,6 +81,9 @@ func New(config Config) (*cobra.Command, error) {
 	}
 
 	f.Init(c)
+
+	c.AddCommand(archiveCmd)
+	c.AddCommand(createCmd)
 
 	return c, nil
 }
